new implementation of authorization using policy

+++ app/Http/Controllers/QuestionsController.php
@@ -6,9 +6,12 @@ use App\Http\Requests\AskQuestionRequest;
 use App\Question;
 use Illuminate\Http\Request;
 use App\User;
-
+use App\Policies\QuestionPolicy;
 class QuestionsController extends Controller
 {
+    public function __construct(){
+        $this->middleware('auth',['except'=>'index','show']);
+    }
 
     public function index()
     {
@@ -22,7 +25,7 @@ class QuestionsController extends Controller
 
     public function create()
     {
-        $question=new Question();
+
 
         return view('questions.create');
     }
@@ -36,25 +39,32 @@ class QuestionsController extends Controller
 
     public function show(Question $question)
     {
-        //
+        $question->increment('views');
+
+        return view('questions.show',array('question'=>$question));
     }
 
 
     public function edit(Question $question)
     {
-        return view('questions.edit',array('question'=>$question));
+      $this->authorize("update",$question);
+        return view('questions.edit', array('question' => $question));
+
     }
 
 
     public function update(AskQuestionRequest $request, Question $question)
-    {
+    {   $this->authorize("update",$question);
         $question->update($request->only('title','body'));
         return redirect()->route('questions.index')->with('success',"Question Updated");
     }
 
 
-    public function destroy($id)
-    {   $question=Question::find($id);
+    public function destroy(Question $question)
+    {
+
+
+        $this->authorize('delete',$question);
         $question->delete();
 
         return redirect()->back()->with('success','Question deleted!');
